% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pooling.R
\docType{package}
\name{advanced_pooling}
\alias{advanced_pooling}
\title{Advanced Pooling Methods for Meta-Analysis}
\description{
Implementation of novel pooling methods validated through simulation study.
These methods provide improved coverage and robustness compared to standard
REML estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mahmood789/Pairwise70}
  \item Report bugs at \url{https://github.com/mahmood789/Pairwise70/issues}
}

}
