% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{Pairwise70-package}
\alias{Pairwise70-package}
\alias{Pairwise70}
\title{Pairwise Meta-Analysis Datasets}
\description{
This package contains pairwise meta-analysis datasets extracted from Cochrane systematic reviews.
Each dataset includes study-level effect sizes, sample sizes, and meta-regression covariates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mahmood789/Pairwise70}
  \item Report bugs at \url{https://github.com/mahmood789/Pairwise70/issues}
}

}
\keyword{package}
